fixed overflow
Former-commit-id: 6266a827cfc5d70c3a2860601de1f316cfb581cb

+++ crates/brontes-classifier/src/classifier.rs
@@ -168,7 +168,7 @@ impl Classifier {
         tree
     }
 
-    fn get_coinbase_transfer(&self, builder: Address, action: &Action) -> Option<u64> {
+    fn get_coinbase_transfer(&self, builder: Address, action: &Action) -> Option<u128> {
         match action {
             Action::Call(action) => {
                 if action.to == builder {
